Moved configuration-related perf agent code into its own module
This change extracts the config module so that we don't pollute main.go
with things that can be isolated. This change should reduce cognitive
complexity and improve modularity.
