+++ bin/juxtaposer/go.mod
@@ -5,5 +5,5 @@ go 1.12
 require (
 	github.com/go-sql-driver/mysql v1.4.1
 	github.com/lib/pq v1.1.1
-	gopkg.in/yaml.v3 v3.0.0-20190502103701-55513cacd4ae // indirect
+	gopkg.in/yaml.v3 v3.0.0-20190502103701-55513cacd4ae
 )
